refactor: extract shared SCSS module
